Server-side Memcache binary protocol read-only filter.
This adds a filter for the Memcache binary protocol, which blocks all
requests that might modify the contents of a Memcache server.

Supports clients which pipeline requests, but it must convert "quiet"
requests to their non-quiet counterparts (GETQ -> GET), in order to
correctly track all oustanding requests to the backend.  This alows it
to respond with errors to RW commands (or close the connection) after
all oustanding responses have been flushed to the client.